Add Measurements panel
Add the skeleton measurements panel that includes the legend and a
container for the SVG. Currently acts as a third "big" panel that can
span the full page or be turned into a half panel for grid display.
Legend is hidden when the panel is in the grid display with the tree.

Reworks the panel displays for the tree and map panels in order to fit
the new measurements panel into the grid. When all three panels are
displayed, the map stays in the "full" mode so that there isnt' an extra
white space next to it.