Add D3 SVG to Measurements panel
Only displays measurements that are linked to samples currently visible
on the tree, so each measurement's "strain" value must exactly match
the strain name of the sample in the tree. This ensures that filters of
the tree also filters data of the multiplot.

Since re-drawing the SVG can be pretty slow, I have taken a couple steps
to memoize the data for the SVG and avoid re-rendering/re-drawing
when not needed:

1. Used `lodash.isEqual` for the React Redux selectors equality checks
for Redux states that are objects. The default is a strict `===`
reference equality check that would fail for objects and cause the
panel to re-render every time an action is dispatched.

2. Used `useMemo` to memoize the D3 scale functions. The D3 scales only
need to be updated when the collection or width has changed, so use
memo to preserve it across renders. Memomizing the functions also allow
them to be used as dependencies of `useEffect` hooks without causing
the hooks to re-run every render.

3. Used a custom `useDeepCompareMemo()` function to memoize all the data
needed for drawing the SVG. This is done in anticipation that following
commits will introduce other `useEffect` hooks that will also need
`svgData` as a dependency (e.g. colors will need to be re-applied when
`svgData` changes).