Measurements: fix continuous color-by mean display
When the color-by is a continuous scale, the individual strain
attributes are grouped into ranges for the legend values of the tree.

To match the measurments means to the legend, group the measurements
by the legend value instead of individual attributes. This does mean
we have to loop through all of the visible legend values for each
visible strain in the tree to determine the matching legend value, which
is definitely not ideal.